Never log to stdout
Since the adoption of GNU parallel, the logging has been a bit off.
This is due to the fact that parallel group output by stdout followed
by stderr, and does not interleave it.

Remove all output to stdout, for log clarity. Also, silence a bunch of
noise from tuxmake which users do not care about.

Finally, cleaned up a bunch of whitespace while walking the file. Diff
with "git diff -w". ;-)

Signed-off-by: Björn Töpel <bjorn@rivosinc.com>